+++ shardingsphere-jdbc/shardingsphere-jdbc-spring/shardingsphere-jdbc-core-spring/shardingsphere-jdbc-core-spring-namespace/src/test/resources/META-INF/rdb/shadowNamespace.xml
@@ -30,14 +30,15 @@
                            http://shardingsphere.apache.org/schema/shardingsphere/shadow
                            http://shardingsphere.apache.org/schema/shardingsphere/shadow/shadow.xsd
                            ">
-    <import resource="datasource/dataSource.xml" />
+    <import resource="datasource/dataSource.xml"/>
 
-    <shardingsphere:data-source id="shadowDataSource" data-source-names="dbtbl_0,dbtbl_1">
-        <shardingsphere:shadow-rule shadow-column="shadow">
-            <shadow:mappings>
-                <shadow:mapping product-data-source-name="dbtbl_0" shadow-data-source-name="dbtbl_1"/>
-            </shadow:mappings>
-        </shardingsphere:shadow-rule>
+    <shadow:rule id="shadowRule" column="shadow">
+        <shadow:mappings>
+            <shadow:mapping product-data-source-name="dbtbl_0" shadow-data-source-name="dbtbl_1"/>
+        </shadow:mappings>
+    </shadow:rule>
+
+    <shardingsphere:data-source id="shadowDataSource" data-source-names="dbtbl_0,dbtbl_1" rule-refs="shadowRule">
         <shardingsphere:props>
             <prop key="sql.show">${sql_show}</prop>
         </shardingsphere:props>
